requirements.txt
contourpy==1.1.1
cycler==0.11.0
fonttools==4.42.1
importlib-resources==6.0.1
joblib==1.3.2
kiwisolver==1.4.5
llvmlite==0.41.0
matplotlib==3.8.0
numba==0.58.0
numpy==1.26.0
packaging==23.1
pandas==2.1.0
Pillow==10.0.1
plotly==5.17.0
pynndescent==0.5.10
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
scikit-learn==1.3.0
scipy==1.11.2
seaborn==0.12.2
six==1.16.0
sklearn==0.0.post9
tbb==2021.10.0
tenacity==8.2.3
threadpoolctl==3.2.0
tqdm==4.66.1
tzdata==2023.3
umap-learn==0.5.4
zipp==3.16.2
